refactor UserIcon
