+++ src/components/PersonalDropdown/PersonalDropdown.tsx
@@ -8,7 +8,7 @@ import Loader from 'react-loader-spinner';
 import { toastSuccess } from '~/utils/toastr';
 
 import { Icon } from '~/components/base/atoms/Icon';
-import { UserIcon } from '~/components/Icons/UserIcon';
+import { UserIcon } from '~/components/domain/User/atoms/UserIcon';
 
 import { useLocale } from '~/hooks/useLocale';
